Some refactoring and added basic tests for the marshal/unmarshal functions
